Implemented a tweet detail component along with url and view button that goes to that component to show a single tweet.
This functionality was linked to django and the api functionality moved to its own module. Unauthenticated users trying to perform actions that need them to be logged in are redirected. Handled in react then after building, this is intergrated into django where we remove default authentication we setup earlier while implementing tweet actions in react